Больше фиксов богу фиксов
