+++ src/view/stats-view.js
@@ -1,16 +1,18 @@
-import { StatFilterType } from '../const';
+import { Ranks, StatFilterType } from '../const';
 import { createDurationMinutes } from '../utils/utils';
 import AbstractView from './abstract-view';
 
 const getStatsTemplate = (stats,rank,activeButton) => {
   const {watched,duration,topGenre} = stats;
-
   return `<section class="statistic">
-  <p class="statistic__rank">
+  ${rank !== Ranks.NONE.name ?
+    `<p class="statistic__rank">
     Your rank
     <img class="statistic__img" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
     <span class="statistic__rank-label">${rank}</span>
-  </p>
+    </p>
+    `
+    :''}
 
   <form action="https://echo.htmlacademy.ru/" method="get" class="statistic__filters">
     <p class="statistic__filters-description">Show stats:</p>
